app/core/config.py
import os
from dotenv import load_dotenv

load_dotenv()

class Config:
    # Project
    PROJECT_NAME = "Mara Backend"
    VERSION = "1.0.0"
    DESCRIPTION = "Flask backend with Redis and Celery for book processing"
    API_V1_STR = "/api/v1"

    # Supabase
    SUPABASE_URL = os.getenv('SUPABASE_URL')
    SUPABASE_KEY = os.getenv('SUPABASE_KEY')

    # CORS
    CORS_ORIGINS = ["*"]
    
    # Redis
    REDIS_HOST = os.getenv('REDIS_HOST', 'localhost')
    REDIS_PORT = int(os.getenv('REDIS_PORT', 6379))
    
    # Celery
    CELERY_BROKER_URL = os.getenv('CELERY_BROKER_URL', 'redis://localhost:6379/0')
    CELERY_RESULT_BACKEND = os.getenv('CELERY_RESULT_BACKEND', 'redis://localhost:6379/0')

    # OpenAI
    OPENAI_API_KEY = os.getenv('OPENAI_API_KEY')
    OPENAI_MODEL = os.getenv('OPENAI_MODEL', 'gpt-4o-mini')
    OPENAI_MAX_TOKENS = int(os.getenv('OPENAI_MAX_TOKENS', 1500))
    OPENAI_SYSTEM_PROMPT = """
    You are a helpful assistant that can extract text from images.
    Make good formatting with shared image.
    Markdown-format with additional formatting rules.
    send only the text in markdown format without '```' or '```markdown'
    """

class DevelopmentConfig(Config):
    DEBUG = True
    TESTING = False

class ProductionConfig(Config):
    DEBUG = False
    TESTING = False

class TestingConfig(Config):
    DEBUG = True
    TESTING = True

# Set the active configuration based on environment
config = {
    'development': DevelopmentConfig,
    'production': ProductionConfig,
    'testing': TestingConfig,
    'default': DevelopmentConfig
}

def get_config():
    env = os.getenv('FLASK_ENV', 'default')
    return config[env]