requirements.txt
flask
flask-cors
flask-restx
marshmallow
celery
redis
supabase
python-multipart
pymupdf
python-dotenv
requests
openai